requirements.txt
amqp==2.5.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-bootstrap==0.2.4
django-braces==1.13.0
django-countries==5.4
django-crispy-forms==1.7.2
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-taggit==1.1.0
djangorestframework==3.10.2
easy-thumbnails==2.6
idna==2.8
importlib-metadata==0.19
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
olefile==0.46
Pillow==6.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
